+++ projects/argfit-ui-core/src/lib/types/button.types.ts
@@ -3,3 +3,5 @@ export type AfButtonVariant = 'primary' | 'secondary' | 'ghost' | 'danger';
 export type AfButtonSize = 'sm' | 'md' | 'lg';
 
 export type AfButtonType = 'button' | 'submit' | 'reset';
+
+export type AfButtonIconPosition = 'start' | 'end';